+++ src/main/java/insane96mcp/iguanatweaksreborn/command/ISOCommand.java
@@ -2,10 +2,10 @@ package insane96mcp.iguanatweaksreborn.command;
 
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.arguments.FloatArgumentType;
+import insane96mcp.iguanatweaksreborn.entity.ISOFallingBlockEntity;
 import insane96mcp.iguanatweaksreborn.module.sleeprespawn.tiredness.TirednessHandler;
 import insane96mcp.iguanatweaksreborn.module.world.weather.Foggy;
 import insane96mcp.iguanatweaksreborn.module.world.weather.Weather;
-import insane96mcp.iguanatweaksreborn.network.message.UnfairOneShotActivation;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.commands.Commands;
 import net.minecraft.commands.arguments.EntityArgument;
@@ -16,6 +16,7 @@ import net.minecraft.world.item.Items;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Blocks;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.server.command.EnumArgument;
 
@@ -82,13 +83,8 @@ public class ISOCommand {
                 )
                 .then(Commands.literal("test")
                         .executes(context -> {
-                            /*for (int i = 0; i < 3; i++) {
-                                ITRFallingBlockEntity fallingBlock = new ITRFallingBlockEntity(context.getSource().getLevel(), context.getSource().getPosition().x, context.getSource().getPosition().y + i, context.getSource().getPosition().z, i == 1 ? Blocks.ACACIA_LOG.defaultBlockState : Blocks.ACACIA_LEAVES.defaultBlockState);
-                                context.getSource().getLevel().addFreshEntity(fallingBlock);
-                            }*/
-                            if (context.getSource().getEntity() instanceof ServerPlayer player) {
-                                UnfairOneShotActivation.send(player);
-                            }
+                            ISOFallingBlockEntity fallingBlock = new ISOFallingBlockEntity(context.getSource().getLevel(), context.getSource().getPosition().x, context.getSource().getPosition().y, context.getSource().getPosition().z, Blocks.DIRT.defaultBlockState);
+                            context.getSource().getLevel().addFreshEntity(fallingBlock);
                             return 1;
                         })));
     }